min/max initializer bug fixed in accel.c
* min/max value is not initialized
* Some platforms(Raspberry Pi 3) is still using Python 2
  So python is changed to python3 to specify the version

+++ include/types.h
@@ -5,7 +5,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-typedef uint16_t float16_t;
+typedef float float16_t;
 typedef float float32_t;
 typedef double float64_t;
